rename connect to loop
