+++ commands/lws
@@ -15,7 +15,7 @@ program
 	.command('ag [number]', 'add-group | Adds one or more new groups (sites) to the database. Number is optional (defaults to one).')
 	.command('al [number]', 'add-layout | Adds one or more new layouts (pages) to the database. Number is optional (defaults to one).')
 	.command('ar [number]', 'add-role | Adds one or more new roles to the database. Number is optional (defaults to one).')
-	.command('connect', 'Manage connect portlet')
+	.command('loop', 'Manage loop portlet')
 	.command('aur', 'add-user-role | Assigns a role to a user.  Will create a new role and user if they are not provided.')
 	.command('gu <userInfo>', 'get-user | Gets a user from the database based on userId, screenName, or emailAddress.')
 	.command('gr <roleInfo>', 'get-role | Gets a role from the database based on roleId or roleName.')
